@@ -6,7 +6,7 @@
   "main": "main/main.ts",
   "scripts": {
     "start": "ts-node src/main.ts",
-    "test": "jest --maxWorkers=2",
+    "test": "jest --detectOpenHandles",
     "dev": "npm run dev:watch & npm run dev:run",
     "dev:watch": "nodemon --watch '*/**.ts' -e ts --exec tsc",
     "dev:run": "nodemon --watch dist/main.js --exec 'node dist/main.js'",
@@ -16,33 +16,42 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@types/ioredis": "^4.28.10",
-    "@types/ioredis-mock": "^5.6.0",
     "@types/jest": "^28.1.6",
-    "@types/node": "^18.0.3",
+    "@types/node": "^18.0.6",
     "@typescript-eslint/eslint-plugin": "^5.30.5",
     "@typescript-eslint/parser": "^5.30.5",
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "eslint-config-airbnb-base": "^15.0.0",
     "eslint-plugin-import": "^2.26.0",
     "ioredis-mock": "^8.2.2",
     "jest": "^28.1.3",
     "nodemon": "^2.0.19",
-    "ts-jest": "^28.0.6",
-    "ts-node": "^10.8.2",
+    "ts-jest": "^28.0.7",
+    "ts-node": "^10.9.1",
     "typescript": "^4.7.4"
   },
   "dependencies": {
-    "@twurple/api": "^5.1.6",
-    "@twurple/auth": "^5.1.6",
-    "@twurple/chat": "^5.1.6",
-    "@twurple/eventsub": "^5.1.6",
-    "@twurple/pubsub": "^5.1.6",
+    "@twurple/api": "^5.1.7",
+    "@twurple/auth": "^5.1.7",
+    "@twurple/chat": "^5.1.7",
+    "@twurple/eventsub": "^5.1.7",
+    "@twurple/pubsub": "^5.1.7",
     "dotenv": "^16.0.1",
     "express": "^4.18.1",
-    "ioredis": "^5.1.0"
+    "ioredis": "^5.2.1"
   },
   "jest": {
+    "roots": [
+      "<rootDir>"
+    ],
+    "modulePaths": [
+      "<rootDir>"
+    ],
+    "moduleDirectories": [
+      "node_modules",
+      "src"
+    ],
+    "testPathIgnorePatterns": ["<rootDir>/build/", "<rootDir>/dist/"],
     "testEnvironment": "node",
     "preset": "ts-jest",
     "collectCoverageFrom": ["./src/**/*.ts"],
@@ -52,6 +61,10 @@
         "branch": 100,
         "function": 100
       }
-    }
+    },
+    "moduleNameMapper": {
+      "^src/(.*)": "<rootDir>/src/$1"
+    },
+    "extensionsToTreatAsEsm": [".spec.ts"]
   }
 }
